Improve NoteRangeRestrictor; reduce size of Converter
The code in Converter that was removed didn't make logical standalone
sense.  The rewrite of NoteRangeRestrictor eliminated the need for it.
In particular, NoteRangeRestrictor now allows operations on ranges of
pitches and durations, instead of just being a container for the lowest
and highest allowable values of each.
This logic improvement nicely reduced the amount of code in many of the
*NoiseGenerator classes.  It should be reasonably straightforward now
also to have NoteRangeRestrictor operate on ranges besides chromatic
sequences of notes (diatonic, etc.)

There are still two classes (FractalNoiseGenerator,
MultiplicativeNoteGenerator) that have not been modified to compile
given the changes.  It appears they are not in active use in the
project.