GUI refactoring
Replaced pitch/length returning functions in SharedPanel with one that
just returns a NoteRangeRestrictor; this reduces the amount of work in
MainPanel, and removes its dependencies on MIDIPitch and Duration.

Instead of RadioGeneratorPanel containing all of the code for each of
the radio buttons and their associated buttons, each radio button panel
is now in its own subclass of a new MusicPanel class.