Removal of Applet entry; additional generated windows; improved layout.
The Applet is no longer considered useful because of the usability of
Java Web Start.
A new MainGuiFrame will be created every time music is generated.  This
is currently automatic, but ideally will be controlled by another
option in the user interface.
The MainPanel utilizes a SpringLayout on one of the child elements to
ensure that it renders as desired.