Minor tweaks / bug fixes
Corrected an error in creation of GridLayout; it did not say to do what
I intended it to if more components were to be added in the future.

Instead of catching any Exception, the code in MainPanel for creating a
MIDI sequence catches particular exceptions and displays slightly
differentiated messages accordingly.

MIDISequenceCreator will now immediately stop playing MIDI notes when
it is halted.  A quick start + stop exposed the problem with *not*
doing this in previous versions; music would continue to play until the
sequencer had run out of data.

MIDISequenceCreator outputs more useful information to standard error
while running.