High and low combo box pairs are aware of each other and only display acceptable values
A bug would allow the "high" value for pitch or duration to be lower
than the "low" value, which would end up causing an exception.
The solution is that the combo boxes now only show values that the user
can select, hiding any that are not acceptable given the value of the
paired combo box.  Every time the selection is changed, the other combo
box in the pair adjusts the values it displays accordingly.