Volume slider added; stop playing music on window close; interface display fix.
There is now a volume slider to control all notes generated from a
particular window.
When a given window is closed, the music being generated from it is
stopped now too.
JComboBoxes that were formerly rendered too narrowly are now widened so
the values they contain can be displayed properly.